更新求职者筛选、状态修改、黑名单、公告等功能

+++ hr-api/src/main/java/com/txxw/hr/service/IAdminService.java
@@ -50,7 +50,7 @@ public interface IAdminService extends IService<Admin> {
      * @param adminId
      * @return
      */
-    Result updateAdminPassword(String oldPass, String pass, Integer adminId);
+    Result updateAdminPassword(String oldPass, String pass, Long adminId);
 
     /**
      * 更新用户头像
